[#17] Get 'mentions' for twitter entities.
This is an accumulative stat that adds the retrieved mentions for an
entity to the mentions value from the latest row in the datastore.

Mention counts are only updated for whole days:

- If the last run was today, running it again will return the value from
  the last run (twitter api isn't used)
- If the last run was before today, the mention count from the date of
  the last run until yesterday is retrieved from the twitter api, and
  added to the value of the last run.
- If there is no value at all (a first run), a value is retrived from a
  few days previously, until yesterday.

This commit also contains a datastore package with a single SQLDatastore
class. This is the interface for processors that need to retrieve data
from a datastore.