[#23] Add a code-packaging pipelne and npm processor
The npm processor fetches `downloads` for each package listed in the
source spec list.

It will fetch for each day from a start date that is either:

- the creation date of the package on npm
- or, the date from the last entry in the db for this package

whichever is most recent.

This allows the processor to collect complete history for the package if
no existing data is present in the db.